requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
django-environ==0.10.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
kombu==5.3.1
Pillow==9.5.0
prompt-toolkit==3.0.38
python-dateutil==2.8.2
redis==4.5.4
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
